refactor: function--saveNewTabToFireStore
