+++ src/utils/firestore.ts
@@ -7,31 +7,6 @@ import {
   updateDoc,
 } from "firebase/firestore";
 import { db } from "../../firebase-config";
-import { getFaviconUrl } from "./tabs";
-
-async function saveTabInfo(tab: chrome.tabs.Tab, userId?: string) {
-  if (!userId || !tab.url || !tab.title || !tab.id) return;
-
-  const tabData = {
-    windowId: tab.windowId,
-    tabId: tab.id,
-    title: tab.title,
-    url: tab.url,
-    favIconUrl: getFaviconUrl(tab.url) || tab.favIconUrl || "",
-    isPinned: false,
-    spaceId: "global",
-  };
-  const newTabId = tab.id;
-  const tabDocRef = doc(db, "users", userId, "tabs", tab.id.toString());
-  const tabOrderDocRef = doc(db, "users", userId, "tabOrders", "global");
-  await setDoc(
-    tabOrderDocRef,
-    { tabOrder: arrayUnion(newTabId), windowId: tab.windowId },
-    { merge: true },
-  );
-  await setDoc(tabDocRef, tabData, { merge: true });
-  return tabData;
-}
 
 async function updateOldTabOrderDoc(
   userId: string,
@@ -72,9 +47,4 @@ async function updateSpaceOfTab(
   }
 }
 
-export {
-  saveTabInfo,
-  updateOldTabOrderDoc,
-  updateNewTabOrderDoc,
-  updateSpaceOfTab,
-};
+export { updateOldTabOrderDoc, updateNewTabOrderDoc, updateSpaceOfTab };
